move ddl generator to pipeline
